refactor: split out variegation functionality
